first

+++ HAM TRONG NGON NGU LAP TRINH C/BAI 1.cpp
@@ -5,12 +5,12 @@ int main(){
 	//BAI 1 : TONG TAT CA CAC CANH CUA HINH HOP CHU NHAT
 	int S1 , S2 , S3; // DIEN TICH BA MAT CO CHUNG DINH
 	scanf("%d %d %d" , &S1 , &S2 , &S3);
-	double sum = 0;// TONG 12 CANH
+	int sum = 0;// TONG 12 CANH
 	double a = 0 , b = 0 , c = 0 ; // ba canh
 	a = sqrt(1.0 * S1 * S3 / S2);
 	b = sqrt(1.0 * S1 * S2 / S3);
 	c = sqrt(1.0 * S2 * S3 / S1);
 	sum = 1.0 * (4 * (a + b + c));
-	printf("%.6lf" , sum);
+	printf("%d" , sum);
 	return 0;
 }
